Generate demotivational quotes for personal use.

demotivate() is a tool to demotivate even the highly motivated individual. Use with caution and remember to not take it personally.

Usage

demotivate(cat = "any", seed)

Arguments

cat

a demotivational category to choose from. The function defaults to 'any'.

seed

a numeric or character seed for reproducible results. The function defaults to 'NULL'.

Value

demotivate() takes a preferred category and an optional reproducible seed to return a demotivational quote for the user to feel worse at any given time.

Details

When cat is 'any' (default), the function combines all category data frames and then samples from the new,combined data frame.

There are a total of six categories: family, life, work, science, tv, and rednecks. When a specific category is chosen the function will sample from a demotivational quote data frame and output a random quote for your spiritual and mental decay.

Author

JP Monteagudo

Examples

demotivate("any",seed = "lanky puppy")
#> [1] "Behind every great man is a woman rolling her eyes. - Bruce Almighty"
demotivate("work")
#> [1] "Work: because someone has to pay for all the coffee you drink."