General motivation quotes for personal use

motivate() is a tool to lift up even the unmotivated individual. Use with caution and remember— DO take it personally.

Usage

motivate(cat = "any", seed)

Arguments

cat

a demotivational category to choose from. The function defaults to 'any'.

seed

a numeric or character seed for reproducible results. The function defaults to 'NULL'.

Value

motivate() takes a preferred category and an optional reproducible seed to return a demotivational quote for the user to feel worse at any given time.

Details

When cat is 'any' (default), the function combines all category data frames and then samples from the new,combined data frame.

There are a total of five categories: psych, stoic, religion, literature(lit), and modern (history) When a specific category is chosen the function will sample from a motivational quote data frame and output a random quote for your spiritual and mental growth.

Author

JP Monteagudo

Examples

motivate("any",seed = "lethargic raccoon")
#> [1] "The world says: 'You have needs -- satisfy them. You have as much right as the rich and the mighty. Don't hesitate to satisfy your needs; indeed, expand your needs and demand more.' This is the worldly doctrine of today. And they believe that this is freedom. The result for the rich is isolation and suicide, for the poor, envy and murder. — Fyodor Dostoevsky"
motivate("lit")
#> [1] "The mystery of human existence lies not in just staying alive, but in finding something to live for. - Fyodor Dostoevsky"
motivate("stoic", seed = 2899)
#> [1] "Wealth consists not in having great possessions, but in having few wants. - Epictetus"