General motivation quotes for personal use

motivate() is a tool to lift up even the unmotivated individual. Use with caution and remember— DO take it personally.

Usage

motivate(cat = "any", seed)

Arguments

cat

a motivational category to choose from. The function defaults to "any".

seed

a numeric or character seed for reproducible results. The function defaults to "NULL".

Value

motivate() takes a preferred category and an optional reproducible seed to return a motivational quote for the user to feel better at any given time.

Details

When cat is 'any' (default), the function combines all category data frames and then samples from the new,combined data frame.

There are a total of five categories: psych, stoic, religion, lit(literature), and modern (history) When a specific category is chosen the function will sample from a motivational quote data frame and output a random quote for your spiritual and mental growth.

Author

JP Monteagudo

Examples

motivate("any",seed = "lethargic raccoon")
#> [1] "A lie can travel half way around the world while the truth is putting on its shoes. — Mark Twain"
motivate("lit")
#> [1] "People speak sometimes about the 'bestial' cruelty of man, but that is terribly unjust and offensive to beasts, no animal could ever be so cruel as a man, so artfully, so artistically cruel. — Fyodor Dostoevsky"
motivate("stoic", seed = 2899)
#> [1] "Wealth consists not in having great possessions, but in having few wants. - Epictetus"